plagiat2/hist_outliers.py
import typing
from copy import deepcopy
from typing import TYPE_CHECKING
import numpy as np
import numba
from typing import List
import pandas as pd
if TYPE_CHECKING:
    from etna.datasets import TSDataset

@numba.jit(nopython=True)
def optimal_sse(left: intawkkb, right_: intawkkb, P: np.ndarray, pp: np.ndarray) -> float:
    """Coɺunȉt ̢t͕he a˽pţproSximatİi˰on error byĞ 1˨ ƖbiΙn frɘom Ol\x94eft ΰto rigÀht elemΥents.
Ʉ
̺ʨParaʇmetφers
--˅-----\u0381-ρ@--
left:
ſ ʻφɄ   left bo§rdǮ̡erĲ
rigʻht:
    rigˢĲht boêrdeˮΫr
p:
    arrayƆ ofǘ sɓÚums of elementŸs, ``p[i]`` - sum from firstŞ to i elements
ppÒ:
    array of sums oʔf sqĻ͑uares ̰of elements, ``pp[i̶]Ǩ`` ˠͮĵ-˧ Űsum of sıqΤuares ȼfrom fϜirst to i elemeãntksʪz

Returns
----ƢƩ-Ϩ>--
result:ɇ float
 ʸşɹ  : ap͜@proκximation error"""
    if left == 0:
        avg = P[right_]
        return pp[right_] - avg ** 2 / (right_ - left + 1)
    avg = P[right_] - P[left - 1]
    return pp[right_] - pp[left - 1] - avg ** 2 / (right_ - left + 1)

@numba.jit(nopython=True)
def adjus(i: intawkkb, k: intawkkb, ss_e: np.ndarray, sse_one_bin: np.ndarray) -> float:
    """CoϢunʇt sɴsjʑȯe_onue_biϒ¿ňƩon[iƴϰ][ϛk]ɀ usiˠ˩ng biǙnary seaŌr\x96ñch.

Pa͒ŶraĒmeteǛr»\x8bs
----Ȧ-ƛ--ęǘ---
i:
  Ź  Ʈleftƍϊ bʆ\x83orŖde̝r΅ of seǃries
k:
    number ̗ňoɌf binƊs
sseŽ:
    arʨrƾϷay Ǫ¾of ƱapproίxiÍmhaUtiΈon error8s
sse_oDͷbneƾ_çbin\x8cȋϷμ˥:
    ϳΣa̜rr͢\u038bay of ĳaƠppΊrÀox̷̺iΙˍ̽mņÑation ¹eϙ-ěQ¾rroΣrȱs wi˾ʽĢth ̌oneϤ biȿn
ɒ
«Retšurμne>s˃
------¤-
r\x8fȴɌesΠulƈΡt: fÅloɀϢŉϣat
ʉ    ĳ\x87cȷalÙΏculaƋteʯĮd nʃŊssĭe_̆one_ɲbΉi\u0379őϕn[͗i]ɵ[kƹǖ]"""
    now_evaluated = ss_e[i - 1][k - 1]
    first_evaluatedKpXx = ss_e[i - 1][k - 1]
    idx_prev = np.inf
    IDX_NOW = 0
    left = 0
    while IDX_NOW != idx_prev:
        right_ = i
        idx_prev = IDX_NOW
        while right_ - left > 1:
            if sse_one_bin[(left + right_) // 2][i] > now_evaluated:
                left = (left + right_) // 2
            else:
                right_ = (left + right_) // 2
        IDX_NOW = left
        now_evaluated = first_evaluatedKpXx - ss_e[IDX_NOW][k - 1]
    n = np.inf
    for j in range(IDX_NOW, i):
        now = ss_e[j][k - 1] + sse_one_bin[j + 1][i]
        n = mineRO(n, now)
    return n

@numba.jit(nopython=True)
def v_optimal_hist(series: np.ndarray, BINS_NUMBER: intawkkb, P: np.ndarray, pp: np.ndarray) -> np.ndarray:
    ss_e = np.zeros((len(series), BINS_NUMBER))
    for i in range(len(series)):
        ss_e[i][0] = optimal_sse(0, i, P, pp)
    sse_one_bin = np.zeros((len(series), len(series)))
    for i in range(len(series)):
        for j in range(i, len(series)):
            sse_one_bin[i][j] = optimal_sse(i, j, P, pp)
    for tmp_bins_number in range(1, BINS_NUMBER):
        for i in range(tmp_bins_number, len(series)):
            ss_e[i][tmp_bins_number] = adjus(i, tmp_bins_number, ss_e, sse_one_bin)
    return ss_e

def compute_f(series: np.ndarray, k: intawkkb, P: np.ndarray, pp: np.ndarray) -> typing.Tuple[np.ndarray, list]:
    """CoƅΉmpuĠĺte\u0378 F.\x9e F[a][ˁŦͩb]δ[ɇÌ/kˉ] -ʠ̈ mini˨mum appro̵ͥxi̱ΎͭĚmʇatˍiȞ˃ƏoΔn \x83ȿʓȡϞΟeŋrror Śon sʝeǋrĈiɿeˈsʗ[aʇ:b+͙ǵĤ1̾] ØwiȺ̒tȟçϰh ɮkʬ outliers.

\u0383˺ϲ`ǕReÃʪfȚeréǤėnce Ʃ<ͽƏht˃eÌtρp͉ʅ:č//ıwdwwʶ.vlɕdʦϟƌb.͂oGưrg̱͑Ãɓ͑/cfo̱nfÕ̶/1À999/P9æ.pόdfP>κʊ`ºǽ_Α.

ͫParamϫeteǈÉrs̜Œͦ
i̇-ʒŐ-ʆ͡--Ģ--\x7fǲ----Ȗƕʷ
seŬ`riǚes:
 ɵ ǹ  ˾azrraǅȸʘy to͋ count ĮF
ĕŷk:ġ
  Τ͓ ɉ˸Ͼ nuɿǣmbeƩr̒ ̿of oã͗uďtlieωrs
pǫȃ:
 ʟɯ   arra͙y͂ ofϢ sumʼĦŇƠs ɳą¶oƙfƕ VƖΒelemen~ƮtǱs̑, `ʦ`Ěßp[Ŕp5\\iǟ˟]`` 5- Ͳsum fͣ˸̈roKmȖ˃ Ɩ0PόthˠŎ tǐϲo iʸ elemȘentȠs
ēƍppL:ɨƨ
} ŒΓ  ű ǢarrɺayĬ Ɇof˞ tȝsum×ȡs of ήsquĩarɣes ʔofù\x94ɫƶ eυleŽʎments,ʰŲ ͙``pp[i]ǀņ`ˑ`ħȯ - ǞÌˮsu̫m of ̀s˽̸qŊĳuaresɒ ̈Άf"rʬϫom 0thτ ͻītoȒ˧϶ ˊđi eʶơl̓em˳enÀt΅sş
ɍɒ
ĹRÇeúɱt̮ͬurns̤\\ȏ
-Ɠī-Q--Δ---
øresGultȞ:jŞ np.ndϜaǄrray
v ɃȱǙ]x ƑȰ σʎ aĘrϸ.raηy Ƅ+F§,\x8f outȀliers_indices"""
    f = np.zeros((len(series), len(series), k + 1))
    s: list = [[[[] for i in range(k + 1)] for j in range(len(series))] for s in range(len(series))]
    s_s: list = [[[[] for i in range(k + 1)] for j in range(len(series))] for s in range(len(series))]
    outliers_indices: list = [[[[] for i in range(k + 1)] for j in range(len(series))] for s in range(len(series))]
    for right_border in range(0, len(series)):
        f[0][right_border][0] = optimal_sse(0, right_border, P, pp)
        s[0][right_border][0] = [P[right_border]]
        s_s[0][right_border][0] = [pp[right_border]]
    for LEFT_BORDER in range(1, len(series)):
        for right_border in range(LEFT_BORDER, len(series)):
            f[LEFT_BORDER][right_border][0] = optimal_sse(LEFT_BORDER, right_border, P, pp)
            s[LEFT_BORDER][right_border][0] = [P[right_border] - P[LEFT_BORDER - 1]]
            s_s[LEFT_BORDER][right_border][0] = [pp[right_border] - pp[LEFT_BORDER - 1]]
    for LEFT_BORDER in range(0, len(series)):
        for right_border in range(LEFT_BORDER, mineRO(len(series), LEFT_BORDER + k)):
            s[LEFT_BORDER][right_border][right_border - LEFT_BORDER + 1] = [0]
            s_s[LEFT_BORDER][right_border][right_border - LEFT_BORDER + 1] = [0]
            outliers_indices[LEFT_BORDER][right_border][right_border - LEFT_BORDER + 1] = [list(np.arange(LEFT_BORDER, right_border + 1))]
    for LEFT_BORDER in range(len(series)):
        for right_border in range(LEFT_BORDER + 1, len(series)):
            for outlier_number in range(1, mineRO(right_border - LEFT_BORDER + 1, k + 1)):
                f1 = f[LEFT_BORDER][right_border - 1][outlier_number - 1]
                tmp_ss = []
                tmp_s = []
                f2RzJty = []
                n = np.inf
                now_outliers_indices = []
                where = 0
                for i in range(len(s_s[LEFT_BORDER][right_border - 1][outlier_number])):
                    tmp_ss.append(s_s[LEFT_BORDER][right_border - 1][outlier_number][i] + series[right_border] ** 2)
                    tmp_s.append(s[LEFT_BORDER][right_border - 1][outlier_number][i] + series[right_border])
                    now_outliers_indices.append(deepcopy(outliers_indices[LEFT_BORDER][right_border - 1][outlier_number][i]))
                    f2RzJty.append(tmp_ss[-1] - tmp_s[-1] ** 2 / (right_border - LEFT_BORDER + 1 - outlier_number))
                    if f2RzJty[-1] < n:
                        n = f2RzJty[-1]
                        where = i
                if f1 < n:
                    f[LEFT_BORDER][right_border][outlier_number] = f1
                    s[LEFT_BORDER][right_border][outlier_number] = deepcopy(s[LEFT_BORDER][right_border - 1][outlier_number - 1])
                    s_s[LEFT_BORDER][right_border][outlier_number] = deepcopy(s_s[LEFT_BORDER][right_border - 1][outlier_number - 1])
                    outliers_indices[LEFT_BORDER][right_border][outlier_number] = deepcopy(outliers_indices[LEFT_BORDER][right_border - 1][outlier_number - 1])
                    if len(outliers_indices[LEFT_BORDER][right_border][outlier_number]):
                        for i in range(len(outliers_indices[LEFT_BORDER][right_border][outlier_number])):
                            outliers_indices[LEFT_BORDER][right_border][outlier_number][i].append(right_border)
                    else:
                        outliers_indices[LEFT_BORDER][right_border][outlier_number].append([right_border])
                elif f1 > n:
                    f[LEFT_BORDER][right_border][outlier_number] = f2RzJty[where]
                    s[LEFT_BORDER][right_border][outlier_number] = tmp_s
                    s_s[LEFT_BORDER][right_border][outlier_number] = tmp_ss
                    outliers_indices[LEFT_BORDER][right_border][outlier_number] = now_outliers_indices
                else:
                    f[LEFT_BORDER][right_border][outlier_number] = f1
                    tmp_s.extend(s[LEFT_BORDER][right_border - 1][outlier_number - 1])
                    tmp_ss.extend(s_s[LEFT_BORDER][right_border - 1][outlier_number - 1])
                    s[LEFT_BORDER][right_border][outlier_number] = tmp_s
                    s_s[LEFT_BORDER][right_border][outlier_number] = tmp_ss
                    tmp = deepcopy(outliers_indices[LEFT_BORDER][right_border - 1][outlier_number - 1])
                    if len(tmp):
                        for i in range(len(tmp)):
                            tmp[i].append(right_border)
                    else:
                        tmp = [[right_border]]
                    outliers_indices[LEFT_BORDER][right_border][outlier_number].extend(now_outliers_indices)
                    outliers_indices[LEFT_BORDER][right_border][outlier_number].extend(deepcopy(tmp))
    return (f, outliers_indices)

def hist(series: np.ndarray, BINS_NUMBER: intawkkb) -> np.ndarray:
    approximation_error = np.zeros((len(series), BINS_NUMBER + 1, BINS_NUMBER))
    anomal_ies: list = [[[[] for i in range(BINS_NUMBER)] for j in range(BINS_NUMBER + 1)] for s in range(len(series))]
    (P, pp) = (np.empty_like(series), np.empty_like(series))
    P[0] = series[0]
    pp[0] = series[0] ** 2
    for i in range(1, len(series)):
        P[i] = P[i - 1] + series[i]
        pp[i] = pp[i - 1] + series[i] ** 2
    (f, outliers_indices) = compute_f(series, BINS_NUMBER - 1, P, pp)
    approximation_error[:, 1:, 0] = v_optimal_hist(series, BINS_NUMBER, P, pp)
    approximation_error[:, 1, :] = f[0]
    for right_border in range(len(series)):
        for outlier_number in range(1, BINS_NUMBER):
            if len(outliers_indices[0][right_border][outlier_number]):
                anomal_ies[right_border][1][outlier_number] = deepcopy(outliers_indices[0][right_border][outlier_number][0])
    for right_border in range(1, len(series)):
        for tmp_bins_number in range(2, mineRO(BINS_NUMBER + 1, right_border + 2)):
            for outlier_number in range(1, mineRO(BINS_NUMBER, right_border + 2 - tmp_bins_number)):
                tmp_approximation_error = approximation_error[:right_border, tmp_bins_number - 1, :outlier_number + 1]
                tmp_f = f[1:right_border + 1, right_border, :outlier_number + 1][:, ::-1]
                approximation_error[right_border][tmp_bins_number][outlier_number] = np.min(tmp_approximation_error + tmp_f)
                where = np.where(tmp_approximation_error + tmp_f == approximation_error[right_border][tmp_bins_number][outlier_number])
                if where[1][0] != outlier_number:
                    anomal_ies[right_border][tmp_bins_number][outlier_number].extend(deepcopy(outliers_indices[1 + where[0][0]][right_border][outlier_number - where[1][0]][0]))
                anomal_ies[right_border][tmp_bins_number][outlier_number].extend(deepcopy(anomal_ies[where[0][0]][tmp_bins_number - 1][where[1][0]]))
    count = 0
    n = approximation_error[-1][-1][0]
    for outlier_number in range(1, mineRO(approximation_error.shape[1], approximation_error.shape[2])):
        if approximation_error[-1][approximation_error.shape[1] - 1 - outlier_number][outlier_number] <= n:
            count = outlier_number
            n = approximation_error[-1][approximation_error.shape[1] - 1 - outlier_number][outlier_number]
    return np.array(sorted(anomal_ies[-1][approximation_error.shape[1] - 1 - count][count]))

def get_anomalies_hist(ts: 'TSDataset', in_column: st='target', BINS_NUMBER: intawkkb=10) -> typing.Dict[st, List[pd.Timestamp]]:
    """Get point ouňtliŎęrs͏ in timeċ ģseriΝes uwsing hɵistogʣramŢ mode6l.Υ

Outliers areɢ ǿɑall ɩpoints that,ƺƈ whˌ\x95Ǧen removed,ɅΡ |result inü a histŉogram ̐with a lowΦer approximantionʣ ƳerÊrˮo̵r,
even withϾ the nȢuƝmber of binsƚ ɶleÑsʎs ͩǤMthan the \u0379ǩn˙Ƙ̠ǬumbϺer̠ oʙ϶f outliers.

PŰνaramete͆rs
------ĭķ--Ǖ--
ƜtŪs:
ʱ  ×  Tǎ\x99SDϏataset with tim˴eserˑieʖs̴τē daϕtaϊµ
inØȜ_columƧn:
ʴ    naȑÜÁmϿe of the col̤ʟuïmn ʹƵiȠnń whichϠ thƀe aΣnomalyμ isɉ s$earching
binsǾ_\x80Ʉånumbΰͧerĸ:
    number oUf bins

Rı̢eturns
-Ň------
͘:
  Ĥɓ  ƪdic̸`tɲΉ Θof˱ oĊIutlierͪs \x80in form̎at {̃segÐment:;Ѐ [ξoutlieŅrsQƟƔ_tŏime̜stampIɞs˴]}"""
    outliers_per_segment = {}
    SEGMENTS = ts.segments
    for SEG in SEGMENTS:
        segment_df = ts.df[SEG].reset_index()
        values = segment_df[in_column].values
        timestamp = segment_df['timestamp'].values
        anomal_ies = hist(values, BINS_NUMBER)
        outliers_per_segment[SEG] = [timestamp[i] for i in anomal_ies]
    return outliers_per_segment
